Extended simplify's resource loading
